Move service registration to async_setup in rainmachine
