+++ homeassistant/components/rainmachine/switch.py
@@ -18,7 +18,7 @@ from homeassistant.helpers import config_validation as cv, entity_platform
 from homeassistant.helpers.entity_platform import AddConfigEntryEntitiesCallback
 from homeassistant.helpers.typing import VolDictType
 
-from . import RainMachineConfigEntry, RainMachineData, async_update_programs_and_zones
+from . import RainMachineConfigEntry, RainMachineData
 from .const import (
     CONF_ALLOW_INACTIVE_ZONES_TO_RUN,
     CONF_DEFAULT_ZONE_RUN_TIME,
@@ -31,6 +31,7 @@ from .const import (
     DEFAULT_ZONE_RUN,
 )
 from .entity import RainMachineEntity, RainMachineEntityDescription
+from .services import async_update_programs_and_zones
 from .util import RUN_STATE_MAP, key_exists
 
 ATTR_ACTIVITY_TYPE = "activity_type"
